First update to rewrite.
The library isn't supposed to use Activities anymore. Everything doesn't work yet.
